Production/backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
mssql-django>=1.3
pyodbc>=5.0
bcrypt>=4.0
<<<<<<< HEAD
<<<<<<< HEAD
python-dotenv>=1.0.0
=======
>>>>>>> a9a5fba (Smart Log inti)
=======
>>>>>>> a9a5fba (Smart Log inti)
